write test for decorators
