+++ capitalize_decorator.rb
@@ -1,0 +1,8 @@
+require_relative 'base_decorator'
+
+class CapitalizeDecorator < BaseDecorator
+  def correct_name
+    capitalized_words = @object.correct_name.chars.map(&:capitalize)
+    capitalized_words.join
+  end
+end
